Login and Signup system
The home_page.html would have a welcome and two buttons. One is register which leads to register.html and login which leads to login.html.
The register.html is asking user's email and password info and has two buttons register which would create user with provided information and login button which lead the user to the login page in case they already registered.
The login.html is asking user to provide email and password and has two buttons. One is login which use the info provided to log user in. The other one is register leading user back to register.html.
The login_system.js handles register and login of a user.